+++ spec/project_spec.rb
@@ -77,7 +77,7 @@ describe 'RuboCop Project' do # rubocop:disable RSpec/DescribeClass
         it 'has a valid URL' do
           issues.each do |issue|
             number = issue[:number].gsub(/\D/, '')
-            pattern = %r{^https://github\.com/[/]+/[/]+/(?:issues|pull)/#{number}$} # rubocop:disable LineLength
+            pattern = %r{^https://github\.com/.+/.+/(?:issues|pull)/#{number}$} # rubocop:disable LineLength
             expect(issue[:url]).to match(pattern)
           end
         end