+++ spec/rubocop/cop/rspec/file_path_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-describe RuboCop::Cop::RSpec::FileName, :config do
+describe RuboCop::Cop::RSpec::FilePath, :config do
   subject(:cop) { described_class.new(config) }
   let(:cop_config) { { 'CustomTransform' => { 'FooFoo' => 'foofoo' } } }